Part DAO/MVC Fronted integration - Part 3 (Licencie is OK)

+++ views/categories/homeView.php
@@ -8,7 +8,7 @@
         <?php include './views/components/flashMsg.php'; ?>
         <!-- Records -->
         <div class="table-responsive">
-            <?php if (!empty($categories)) : ?>
+            <?php if (count($categories) > 1) : ?>
                 <table class="table text-nowrap mb-0 align-middle">
                     <thead class="text-dark fs-4">
                         <tr>
@@ -25,19 +25,21 @@
                     </thead>
                     <tbody>
                         <?php foreach ($categories as $categorie) : ?>
-                            <tr>
-                                <td class="border-bottom-0">
-                                    <h6 class="fw-semibold mb-0"><?= $categorie->getNom(); ?></h6>
-                                </td>
-                                <td class="border-bottom-0">
-                                    <h6 class="fw-semibold mb-1"><?= $categorie->getCode(); ?></h6>
-                                </td>
-                                <td class="border-bottom-0 d-flex">
-                                    <a class="btn btn-primary m-1" href="index.php?page=categories&action=show&id=<?= $categorie->getId(); ?>">Voir</a>
-                                    <a class="btn btn-warning m-1" href="index.php?page=categories&action=edit&id=<?= $categorie->getId(); ?>">Modifier</a>
-                                    <a class="btn btn-danger m-1" href="index.php?page=categories&action=delete&id=<?= $categorie->getId(); ?>">Supprimer</a>
-                                </td>
-                            </tr>
+                            <?php if ($categorie->getId() != 1) : ?>
+                                <tr>
+                                    <td class="border-bottom-0">
+                                        <h6 class="fw-semibold mb-0"><?= $categorie->getNom(); ?></h6>
+                                    </td>
+                                    <td class="border-bottom-0">
+                                        <h6 class="fw-semibold mb-1"><?= $categorie->getCode(); ?></h6>
+                                    </td>
+                                    <td class="border-bottom-0 d-flex">
+                                        <a class="btn btn-primary m-1" href="index.php?page=categories&action=show&id=<?= $categorie->getId(); ?>">Voir</a>
+                                        <a class="btn btn-warning m-1" href="index.php?page=categories&action=edit&id=<?= $categorie->getId(); ?>">Modifier</a>
+                                        <a class="btn btn-danger m-1" href="index.php?page=categories&action=delete&id=<?= $categorie->getId(); ?>">Supprimer</a>
+                                    </td>
+                                </tr>
+                            <?php endif; ?>
                         <?php endforeach; ?>
                     </tbody>
                 </table>
